文件缩紧及文件后缀
