+++ src/components/Upload/gallery/pause-icon.js
