+++ src/components/Upload/gallery/upload-failed-icon.js
