+++ src/components/Upload/gallery/upload-icon.js
